Updated to MAME 0.117u1 core
git-svn-id: svn://messdev.no-ip.org/mess@720 9849079f-6229-0410-a07f-84406ce305ba

+++ src/osd/windows/winalloc.c
@@ -345,7 +345,7 @@ void CLIB_DECL free(void *memory)
 
 	// free the memory
 	if (USE_GUARD_PAGES)
-		VirtualFree((UINT8 *)memory - ((UINT32)memory & (PAGE_SIZE-1)) - PAGE_SIZE, 0, MEM_RELEASE);
+		VirtualFree((UINT8 *)memory - ((size_t)memory & (PAGE_SIZE-1)) - PAGE_SIZE, 0, MEM_RELEASE);
 	else
 		GlobalFree(memory);
